+++ test/json_action_order.cpp
@@ -138,3 +138,41 @@ TEST_CASE("create_success_new", "[json_action_order]") {
   };
   ParserTester<value_type>::dispatch(helper, message, 8192, 2);
 }
+
+TEST_CASE("create_failure_20260803", "[json_action_order]") {
+  auto message = R"({)"
+                 R"("channel":"post",)"
+                 R"("data":{)"
+                 R"("id":3000005,)"
+                 R"("response":{)"
+                 R"("type":"action",)"
+                 R"("payload":{)"
+                 R"("status":"ok",)"
+                 R"("response":{)"
+                 R"("type":"order",)"
+                 R"("data":{)"
+                 R"("statuses":[{)"
+                 R"("error":"Order value too large. Max is $5000000.")"
+                 R"(})"
+                 R"(])"
+                 R"(})"
+                 R"(})"
+                 R"(})"
+                 R"(})"
+                 R"(})"
+                 R"(})"sv;
+  auto helper = [](value_type const &obj) {
+    CHECK(obj.channel == protocol::json::Channel::POST);
+    CHECK(obj.data.id == 3000005);
+    CHECK(obj.data.response.type == protocol::json::ResponseType::ACTION);
+    CHECK(obj.data.response.payload.status == protocol::json::Status::OK);
+    CHECK(obj.data.response.payload.response.type == protocol::json::ResponseType::ORDER);
+    //
+    REQUIRE(std::size(obj.data.response.payload.response.data.statuses) == 1);
+    auto &s0 = obj.data.response.payload.response.data.statuses[0];
+    CHECK(s0.error == R"(Order value too large. Max is $5000000.)"sv);
+    CHECK(s0.resting.oid == 0);
+    CHECK(std::empty(s0.resting.cloid));
+  };
+  ParserTester<value_type>::dispatch(helper, message, 8192, 2);
+}